# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Spring Boot API Demo - Programación III
  version: 1.0.0
  description: API contract-first para taller Spring Boot 3.x
  contact:
    name: Workshop Team
    email: workshop@ejemplo.com

servers:
  - url: http://localhost:8080

tags:
  - name: workshop
    description: Endpoints base del workshop
  - name: simulaciones
    description: Simulador de préstamos
  - name: demo-estado
    description: Demo de estado singleton vs manual

paths:
  /api/v1:
    get:
      tags: [workshop]
      operationId: getWorkshopHealth
      summary: Health check del workshop
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/saludos:
    get:
      tags: [workshop]
      operationId: saludarPorGet
      summary: Saludo vía query param
      parameters:
        - name: nombre
          in: query
          required: false
          schema:
            type: string
            default: Mundo
      responses:
        '200':
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
    post:
      tags: [workshop]
      operationId: saludarPorPost
      summary: Saludo vía request body
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaludoRequest'
      responses:
        '200':
          description: Saludo generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaludoResponse'
        '400':
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/simulaciones/prestamo:
    post:
      tags: [simulaciones]
      operationId: simularPrestamo
      summary: Simular cuota de préstamo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PrestamoRequest'
      responses:
        '200':
          description: Simulación exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrestamoResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/demo/estado/singleton/{valor}:
    post:
      tags: [demo-estado]
      operationId: actualizarSingleton
      summary: Actualizar estado singleton
      parameters:
        - name: valor
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Estado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/singleton:
    get:
      tags: [demo-estado]
      operationId: obtenerSingleton
      summary: Obtener estado singleton
      responses:
        '200':
          description: Estado actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/singleton/reset:
    post:
      tags: [demo-estado]
      operationId: reiniciarSingleton
      summary: Reiniciar estado singleton
      responses:
        '200':
          description: Estado reiniciado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/manual/{valor}:
    post:
      tags: [demo-estado]
      operationId: actualizarManual
      summary: Actualizar estado manual
      parameters:
        - name: valor
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Estado temporal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

  /api/v1/demo/estado/manual:
    get:
      tags: [demo-estado]
      operationId: obtenerManual
      summary: Obtener estado manual (nueva instancia)
      responses:
        '200':
          description: Estado manual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoResponse'

components:
  schemas:
    HealthResponse:
      type: object
      required: [estado, mensaje]
      properties:
        estado:
          type: string
          example: ok
        mensaje:
          type: string
          example: Workshop Spring Boot activo

    SaludoRequest:
      type: object
      required: [nombre]
      properties:
        nombre:
          type: string
          minLength: 1
          example: Ana

    SaludoResponse:
      type: object
      required: [mensaje]
      properties:
        mensaje:
          type: string
          example: Hola, Ana!

    PrestamoRequest:
      type: object
      required: [monto, tasaAnual, meses]
      properties:
        monto:
          type: number
          minimum: 0.01
          example: 10000
        tasaAnual:
          type: number
          minimum: 0.01
          example: 12.5
        meses:
          type: integer
          format: int32
          minimum: 1
          maximum: 360
          example: 24

    PrestamoResponse:
      type: object
      required: [cuotaMensual, interesTotal, totalPagar]
      properties:
        cuotaMensual:
          type: number
          example: 470.73
        interesTotal:
          type: number
          example: 1297.52
        totalPagar:
          type: number
          example: 11297.52

    EstadoResponse:
      type: object
      required: [tipo, valorActual]
      properties:
        tipo:
          type: string
          enum: [singleton, manual]
          example: singleton
        valorActual:
          type: integer
          format: int32
          example: 25

    ErrorResponse:
      type: object
      required: [codigo, mensaje, timestamp]
      properties:
        codigo:
          type: string
          example: BUSINESS_RULE_ERROR
        mensaje:
          type: string
          example: Uno o mas campos son invalidos
        timestamp:
          type: string
          format: date-time
          example: 2026-05-15T21:30:00-06:00
        detalles:
          type: object
          additionalProperties:
            type: string
          example:
            nombre: El nombre es obligatorio
            monto: El monto debe ser mayor que 0